# backend/app/api/static/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Track, Manage, and Organize your Shit.",
        "title": "Homebox API",
        "contact": {
            "name": "Don't"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/v1/items": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Query All Items",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search string",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "items per page",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "label Ids",
                        "name": "labels",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "location Ids",
                        "name": "locations",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Create Item",
                "parameters": [
                    {
                        "description": "Item Data",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/repo.ItemCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/repo.ItemSummary"
                        }
                    }
                }
            }
        },
        "/v1/items/export": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Export Items",
                "responses": {
                    "200": {
                        "description": "text/csv",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/items/fields": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Get All Custom Field Names",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/items/fields/values": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Get All Custom Field Values",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/items/import": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Import Items",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Image to upload",
                        "name": "csv",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/v1/items/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Get Item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/repo.ItemOut"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Update Item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Item Data",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/repo.ItemUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/repo.ItemOut"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Delete Item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "repo.DocumentOut": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "repo.ItemAttachment": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "document": {
                    "$ref": "#/definitions/repo.DocumentOut"
                },
                "id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "repo.ItemCreate": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "labelIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "locationId": {
                    "description": "Edges",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "string",
                    "x-nullable": true
                }
            }
        },
        "repo.ItemField": {
            "type": "object",
            "properties": {
                "booleanValue": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "numberValue": {
                    "type": "integer"
                },
                "textValue": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "repo.ItemOut": {
            "type": "object",
            "properties": {
                "archived": {
                    "type": "boolean"
                },
                "assetId": {
                    "type": "string",
                    "example": "0"
                },
                "attachments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repo.ItemAttachment"
                    }
                },
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repo.ItemSummary"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repo.ItemField"
                    }
                },
                "id": {
                    "type": "string"
                },
                "insured": {
                    "type": "boolean"
                },
                "labels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repo.LabelSummary"
                    }
                },
                "lifetimeWarranty": {
                    "description": "Warranty",
                    "type": "boolean"
                },
                "location": {
                    "description": "Edges",
                    "allOf": [
                        {
                            "$ref": "#/definitions/repo.LocationSummary"
                        }
                    ],
                    "x-nullable": true,
                    "x-omitempty": true
                },
                "manufacturer": {
                    "type": "string"
                },
                "modelNumber": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "description": "Extras",
                    "type": "string"
                },
                "parent": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/repo.ItemSummary"
                        }
                    ],
                    "x-nullable": true,
                    "x-omitempty": true
                },
                "purchaseFrom": {
                    "type": "string"
                },
                "purchasePrice": {
                    "type": "string",
                    "example": "0"
                },
                "purchaseTime": {
                    "description": "Purchase",
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "serialNumber": {
                    "type": "string"
                },
                "soldNotes": {
                    "type": "string"
                },
                "soldPrice": {
                    "type": "string",
                    "example": "0"
                },
                "soldTime": {
                    "description": "Sold",
                    "type": "string"
                },
                "soldTo": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "warrantyDetails": {
                    "type": "string"
                },
                "warrantyExpires": {
                    "type": "string"
                }
            }
        },
        "repo.ItemSummary": {
            "type": "object",
            "properties": {
                "archived": {
                    "type": "boolean"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "insured": {
                    "type": "boolean"
                },
                "labels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repo.LabelSummary"
                    }
                },
                "location": {
                    "description": "Edges",
                    "allOf": [
                        {
                            "$ref": "#/definitions/repo.LocationSummary"
                        }
                    ],
                    "x-nullable": true,
                    "x-omitempty": true
                },
                "name": {
                    "type": "string"
                },
                "purchasePrice": {
                    "type": "string",
                    "example": "0"
                },
                "quantity": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "repo.ItemUpdate": {
            "type": "object",
            "properties": {
                "archived": {
                    "type": "boolean"
                },
                "assetId": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repo.ItemField"
                    }
                },
                "id": {
                    "type": "string"
                },
                "insured": {
                    "type": "boolean"
                },
                "labelIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "lifetimeWarranty": {
                    "description": "Warranty",
                    "type": "boolean"
                },
                "locationId": {
                    "description": "Edges",
                    "type": "string"
                },
                "manufacturer": {
                    "type": "string"
                },
                "modelNumber": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "description": "Extras",
                    "type": "string"
                },
                "parentId": {
                    "type": "string",
                    "x-nullable": true,
                    "x-omitempty": true
                },
                "purchaseFrom": {
                    "type": "string"
                },
                "purchasePrice": {
                    "type": "string",
                    "example": "0"
                },
                "purchaseTime": {
                    "description": "Purchase",
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "serialNumber": {
                    "description": "Identifications",
                    "type": "string"
                },
                "soldNotes": {
                    "type": "string"
                },
                "soldPrice": {
                    "type": "string",
                    "example": "0"
                },
                "soldTime": {
                    "description": "Sold",
                    "type": "string"
                },
                "soldTo": {
                    "type": "string"
                },
                "warrantyDetails": {
                    "type": "string"
                },
                "warrantyExpires": {
                    "description": "Sold",
                    "type": "string"
                }
            }
        },
...
        "repo.LabelSummary": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "repo.LocationSummary": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "\"Type 'Bearer TOKEN' to correctly set the API Key\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}